+++ platform/src/main/java/ch/heigvd/amt/gamification/dto/EventCreationDTO.java
@@ -1,10 +1,6 @@
 package ch.heigvd.amt.gamification.dto;
 
-import java.util.LinkedList;
-import java.util.List;
 import com.fasterxml.jackson.annotation.JsonProperty;
-
-import java.math.BigDecimal;
 
 /**
  * EventtypeDTO
@@ -12,24 +8,24 @@ import java.math.BigDecimal;
 public class EventCreationDTO {
 
     @JsonProperty("user_id")
-    private int user_id;
+    private Long user_id;
 
     @JsonProperty("eventtype_id")
-    private long eventtype_id;
+    private Long eventtype_id;
 
-    public int getUser_id() {
+    public Long getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(int user_id) {
+    public void setUser_id(Long user_id) {
         this.user_id = user_id;
     }
 
-    public long getEventtype_id() {
+    public Long getEventtype_id() {
         return eventtype_id;
     }
 
-    public void setEventtype_id(long eventtype_id) {
+    public void setEventtype_id(Long eventtype_id) {
         this.eventtype_id = eventtype_id;
     }
 }
